split up cpp pybind to allow reuse / add simple simulator pybind lib for assessing future changes
