+++ pyphare/pyphare/cpp/__init__.py
@@ -1,13 +1,27 @@
 
+# continue to use override if set
+_cpp_lib_override = None
 
-def cpp_lib():
+def cpp_lib(override=None):
     import importlib
+
+    global _cpp_lib_override
+    if override is not None:
+        _cpp_lib_override = override
+    if _cpp_lib_override is not None:
+        return importlib.import_module(_cpp_lib_override)
+
     if not __debug__:
         return importlib.import_module("pybindlibs.cpp")
     try:
         return importlib.import_module("pybindlibs.cpp_dbg")
     except ImportError as err:
         return importlib.import_module("pybindlibs.cpp")
+
+
+def cpp_etc_lib():
+    import importlib
+    return importlib.import_module("pybindlibs.cpp_etc")
 
 
 def splitter_type(dim, interp, n_particles):
