+++ res/cmake/bench.cmake
@@ -50,5 +50,6 @@ if (bench)
   add_subdirectory(tools/bench/core/numerics/pusher)
 
   add_subdirectory(tools/bench/hi5)
+  add_subdirectory(tools/bench/real)
 
 endif()
