+++ tools/bench/functional/.gitignore
@@ -1,1 +1,1 @@
-uniform
+generated
